apps/datalens-bot/requirements.txt
selenium==4.27.1
Pillow==11.0.0
requests==2.32.3
